Rename aux.js for compatibility with Windows OS.
