+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "wyvern-js",
-  "version": "3.0.2",
+  "version": "3.0.3",
   "description": "Javascript SDK for the Wyvern Protocol",
   "main": "lib/index.js",
   "scripts": {
